refactor(memory): split backfill lifecycle helpers
